move cartBubble to utils.mjs
