+++ src/js/main.js
@@ -1,8 +1,3 @@
 import { loadHeaderFooter } from "./utils.mjs";
-import { displayCartBubble } from "./cartBubble";
 
 loadHeaderFooter();
-
-setTimeout(() => {
-  displayCartBubble();
-}, 20);
