+++ src/js/product-list.js
@@ -1,11 +1,7 @@
 import productList from "./productList.mjs";
-import { getParam, loadHeaderFooter } from "./utils.mjs";
-import { displayCartBubble } from "./cartBubble";
+import { getParam, loadHeaderFooter, displayCartBubble } from "./utils.mjs";
 
 loadHeaderFooter();
-setTimeout(() => {
-  displayCartBubble();
-}, 20);
 
 const category = getParam("category");
 productList(".product-list", category);
