+++ src/js/utils.mjs
@@ -75,6 +75,21 @@ export async function loadTemplate(path) {
   };
 }
 
+export function displayCartBubble() {
+  const existingCart = getLocalStorage("so-cart");
+  const cartLength = existingCart ? existingCart.length : 0;
+  const bubble = document.getElementById("bubble");
+  if (bubble) {
+    if (cartLength > 0) {
+      bubble.textContent = cartLength;
+      bubble.classList.add("cart-bubble");
+    } else {
+      bubble.classList.remove("cart-bubble");
+      document.getElementById("bubble").innerHTML = "";
+    }
+  }
+}
+
 export async function loadHeaderFooter() {
   const headerElement = document.querySelector("#home-header");
   const footerElement = document.querySelector("#home-footer");
@@ -82,4 +97,7 @@ export async function loadHeaderFooter() {
   const footerTemplate = await loadTemplate("/partials/footer.html");
   renderWithTemplate(headerTemplate, headerElement);
   renderWithTemplate(footerTemplate, footerElement);
+  setTimeout(() => {
+    displayCartBubble();
+  }, 50);
 }
